Extract interface from ConfigMerger; impl is DefaultConfigMerger.
