+++ robolectric/src/test/java/org/robolectric/DefaultSdkPickerTest.java
@@ -6,6 +6,9 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import android.os.Build.VERSION_CODES;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -13,19 +16,20 @@ import org.junit.runners.JUnit4;
 import org.robolectric.annotation.Config;
 import org.robolectric.annotation.internal.ConfigUtils;
 import org.robolectric.internal.DefaultSdkProvider;
+import org.robolectric.internal.SdkConfig;
 
 @RunWith(JUnit4.class)
-public class SdkPickerTest {
+public class DefaultSdkPickerTest {
   private static final int[] sdkInts = { 16, 17, 18, 19, 21, 22, 23 };
   private UsesSdk usesSdk;
   private SdkPicker sdkPicker;
-  private DefaultSdkProvider sdkProvider;
+  private SdkProvider sdkProvider;
 
   @Before
   public void setUp() throws Exception {
     usesSdk = mock(UsesSdk.class);
     sdkProvider = new DefaultSdkProvider();
-    sdkPicker = new DefaultSdkPicker(sdkProvider, sdkProvider.map(sdkInts), null);
+    sdkPicker = new DefaultSdkPicker(sdkProvider, map(sdkInts), null);
   }
 
   @Test
@@ -168,9 +172,7 @@ public class SdkPickerTest {
   public void withEnabledSdks_shouldRestrictAsSpecified() throws Exception {
     when(usesSdk.getMinSdkVersion()).thenReturn(16);
     when(usesSdk.getMaxSdkVersion()).thenReturn(23);
-    DefaultSdkProvider sdkProvider = new DefaultSdkProvider();
-    sdkPicker = new DefaultSdkPicker(sdkProvider,
-        sdkProvider.map(sdkInts), sdkProvider.map(17, 18));
+    sdkPicker = new DefaultSdkPicker(sdkProvider, map(sdkInts), map(17, 18));
     assertThat(sdkPicker.selectSdks(new Config.Builder().setSdk(Config.ALL_SDKS).build(), usesSdk))
         .containsExactly(sdkProvider.getSdkConfig(17), sdkProvider.getSdkConfig(18));
   }
@@ -183,5 +185,9 @@ public class SdkPickerTest {
     assertThat(ConfigUtils.parseSdkArrayProperty("KITKAT, LOLLIPOP"))
         .asList()
         .containsExactly(VERSION_CODES.KITKAT, VERSION_CODES.LOLLIPOP);
+  }
+
+  private List<SdkConfig> map(int... sdkInts) {
+    return Arrays.stream(sdkInts).mapToObj(sdkProvider::getSdkConfig).collect(Collectors.toList());
   }
 }
