+++ robolectric/src/test/java/org/robolectric/RobolectricTestRunnerTest.java
@@ -42,7 +42,6 @@ import org.robolectric.annotation.Config;
 import org.robolectric.internal.DefaultSdkProvider;
 import org.robolectric.internal.ParallelUniverseInterface;
 import org.robolectric.internal.SdkEnvironment;
-import org.robolectric.internal.SdkProvider;
 import org.robolectric.manifest.AndroidManifest;
 import org.robolectric.util.Injector;
 import org.robolectric.util.PerfStatsCollector.Metric;
