+++ robolectric/src/test/java/org/robolectric/util/TestUtil.java
@@ -12,9 +12,9 @@ import org.junit.runners.model.InitializationError;
 import org.robolectric.LegacyDependencyResolver;
 import org.robolectric.R;
 import org.robolectric.RobolectricTestRunner;
+import org.robolectric.SdkProvider;
 import org.robolectric.internal.DefaultSdkProvider;
 import org.robolectric.internal.SdkConfig;
-import org.robolectric.internal.SdkProvider;
 import org.robolectric.internal.dependency.DependencyResolver;
 import org.robolectric.res.Fs;
 import org.robolectric.res.ResourcePath;
